+++ src/PSRule.Rules.Azure/rules/Azure.FrontDoor.Rule.ps1
@@ -21,7 +21,6 @@ Rule 'Azure.FrontDoor.MinTLS' -Type 'Microsoft.Network/frontDoors', 'Microsoft.N
     foreach ($endpoint in $endpoints) {
         $Assert.HasDefaultValue($endpoint, 'properties.customHttpsConfiguration.minimumTlsVersion', '1.2');
     }
-    # properties.frontendEndpoints[].properties.customHttpsConfiguration.minimumTlsVersion
 }
 
 # Synopsis: Use diagnostics to audit Front Door access
@@ -54,12 +53,12 @@ Rule 'Azure.FrontDoor.Name' -Type 'Microsoft.Network/frontDoors' -Tag @{ release
     # https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftnetwork
 
     # Between 5 and 64 characters long
-    $Assert.GreaterOrEqual($TargetObject, 'Name', 5)
-    $Assert.LessOrEqual($TargetObject, 'Name', 64)
+    $Assert.GreaterOrEqual($PSRule, 'TargetName', 5);
+    $Assert.LessOrEqual($PSRule, 'TargetName', 64);
 
     # Alphanumerics and hyphens
     # Start and end with alphanumeric
-    Match 'Name' '^[A-Za-z](-|[A-Za-z0-9])*[A-Za-z0-9]$'
+    $Assert.Match($PSRule, 'TargetName', '^[A-Za-z](-|[A-Za-z0-9])*[A-Za-z0-9]$');
 }
 
 # Synopsis: Use Front Door WAF policy in prevention mode
@@ -70,6 +69,17 @@ Rule 'Azure.FrontDoor.WAF.Mode' -Type 'Microsoft.Network/frontdoorwebapplication
 # Synopsis: Enable Front Door WAF policy
 Rule 'Azure.FrontDoor.WAF.Enabled' -Type 'Microsoft.Network/frontdoorwebapplicationfirewallpolicies' -Tag @{ release = 'GA'; ruleSet = '2020_06' } {
     $Assert.HasFieldValue($TargetObject, 'Properties.policySettings.enabledState', 'Enabled');
+}
+
+# Synopsis: Use Front Door WAF naming requirements
+Rule 'Azure.FrontDoor.WAF.Name' -Type 'Microsoft.Network/frontdoorwebapplicationfirewallpolicies' -Tag @{ release = 'GA'; ruleSet = '2020_12' } {
+    # Between 1 and 128 characters long
+    $Assert.GreaterOrEqual($PSRule, 'TargetName', 1);
+    $Assert.LessOrEqual($PSRule, 'TargetName', 128);
+
+    # Letters or numbers
+    # Start letter
+    $Assert.Match($PSRule, 'TargetName', '^[A-Za-z][A-Za-z0-9]{0,127}$');
 }
 
 #endregion Front Door